Connecting the client side auth logic to the server side.
 I put some rudimentary authorization in the routes to try to prevent people
 from making requests on assets that they don't own. Hoever, it feels a little
 clunky and it's incomplete (you can post a step to any arbitrary plan).

 The clunkiness stems from me not wanting to hit the database to get the plan
 and check the accountId, only to hit the database again. So, what I do is fetch
 the data and then check the accountIds. This works but is clunky for the cases
 where there is no data to be fetched (POST) and the flow just seems backwards.
 Why put authorization logic in the middle of completing a task. If I want to
 break the logic out into a mapper function, I'll need to determine auth logic
 beforehand anyway. Just seems like a waste at this point.

+++ routes/plans.js
@@ -3,16 +3,19 @@
 const express = require('express');
 const router = express.Router();
 const Plan = require('../models/plan');
+const AuthorizationError = require('../lib/authorization-error');
+const routeHelper = require('./route-helper');
 
 router.get('/', function allPlansRoute(req, res, next) {
-  Plan.getJoin().run().then(function allPlansDbCallback(plans) {
+  Plan.filter({ accountId: req.user.account_id }).getJoin().run().then(function allPlansDbCallback(plans) {
     res.json(plans);
   }).catch(next);
 });
 
 router.post('/', function postPlanRoute(req, res, next) {
   let plan = new Plan({
-    name: req.body.name
+    name: req.body.name,
+    accountId: req.user.account_id
   });
 
   plan.save().then(function postPlanDbCallback(plan) {
@@ -28,12 +31,20 @@ router.get('/:planId', function getPlanRoute(req, res, next) {
       }
     }
   }).run().then(function getPlanDbCallback(plan) {
+    if (!routeHelper.isAuthorized(req, plan.accountId)) {
+      return next(new AuthorizationError());
+    }
+
     res.json(plan);
   }).catch(next);
 });
 
 router.patch('/:planId', function patchPlanRoute(req, res, next) {
   Plan.get(req.params.planId).run().then(function postPlanDbCallback(plan) {
+    if (!routeHelper.isAuthorized(req, plan.accountId)) {
+      return next(new AuthorizationError());
+    }
+
     plan.merge(req.body).save().then(function savePlanDbCallback(updatedPlan) {
       res.json(updatedPlan);
     }).catch(next);
@@ -42,6 +53,10 @@ router.patch('/:planId', function patchPlanRoute(req, res, next) {
 
 router.delete('/:planId', function deletePlanRoute(req, res, next) {
   Plan.get(req.params.planId).getJoin().then(function getPlanDbCallback(plan) {
+    if (!routeHelper.isAuthorized(req, plan.accountId)) {
+      return next(new AuthorizationError());
+    }
+
     plan.deleteAll().then(function deletePlanDbCallback(deletedPlan) {
       res.json(deletedPlan);
     }).catch(next);
